Menambahkan kolom qty and price di class product

+++ src/main/java/pws/c/restApi_Projects/ProductServiceController.java
@@ -27,11 +27,15 @@ public class ProductServiceController {
         Product honey = new Product();
         honey.setId("1");
         honey.setName("Madu");
+        honey.setPrice("5000");
+        honey.setQty("6");
         productRepo.put(honey.getId(), honey);
         
         Product almond = new Product();
         almond.setId("2");
         almond.setName("almond");
+        almond.setPrice("6000");
+        almond.setQty("7");
         productRepo.put(almond.getId(), almond);
     }
     
@@ -46,7 +50,7 @@ public class ProductServiceController {
     public ResponseEntity<Object> createProduct(@RequestBody Product product){
         
         if(productRepo.containsKey(product.getId())){
-            return new ResponseEntity<>("ID Product Cannot be the Same, please check again", HttpStatus.OK);
+            return new ResponseEntity<>("ID Product "  + product.getId().toString() +" Cannot be the Same" +",Please check again", HttpStatus.OK);
         }
         else{
             productRepo.put(product.getId(), product);
@@ -59,7 +63,7 @@ public class ProductServiceController {
     public ResponseEntity<Object> updateProduct(@PathVariable("id") String id, @RequestBody Product product){
         
         if(!productRepo.containsKey(id)){
-            return new ResponseEntity<>("Product Not Found, Please check again", HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("Product Not Found to updated ID " + id +" ,Please check again", HttpStatus.NOT_FOUND);
         }
         else{
             productRepo.remove(id);
@@ -73,8 +77,16 @@ public class ProductServiceController {
     // DELETE API
     @RequestMapping(value = "/products/{id}", method = RequestMethod.DELETE)
     public ResponseEntity<Object> deleteProduct(@PathVariable("id") String id){
-        productRepo.remove(id);
-        return new ResponseEntity<>("Product is deleted successfully", HttpStatus.OK);
+        
+        if(!productRepo.containsKey(id)){
+            return new ResponseEntity<>("Product Not Found to deleted " + id +" ,Please check again", HttpStatus.NOT_FOUND);
+        }
+        else{
+            productRepo.remove(id);
+            return new ResponseEntity<>("Product is deleted successfully", HttpStatus.OK);
+        }
+        
+        
     }
     
 }
